HTTP 응답 처리를 위한 Response 클래스 리팩토링
HTTP 응답 처리를 위한 코드를 리팩토링하였습니다. 이를 통해 코드의 중복을 줄이고, 확장성을 높였습니다.

주요 변경사항:
1. HttpStatus enum 클래스를 삭제하였습니다. 이 클래스는 더 이상 사용되지 않습니다.
2. Response 클래스를 새로 생성하였습니다. 이 클래스는 HTTP 응답을 처리하는 기본 클래스로 사용됩니다.
3. ResponseNotFound와 ResponseSuccess 클래스가 Response 클래스를 상속받도록 변경하였습니다. 이를 통해 공통된 응답 처리 로직을 재사용할 수 있습니다.
4. RequestHandler 클래스에서 ResponseNotFound의 send 메소드 호출 시 null을 전달하도록 변경하였습니다. 이는 Response 클래스의 send 메소드가 String을 인자로 받도록 변경되었기 때문입니다.
